Переход на annotation-based модель
