+++ src/main/java/com/rntgroup/db/UserDatabase.java
@@ -9,12 +9,14 @@ import lombok.Setter;
 import lombok.experimental.FieldDefaults;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Component;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
+@Component
 @Setter
 @Getter
 @FieldDefaults(level = AccessLevel.PRIVATE)
